requirements.txt
Django==4.2.9
djangorestframework==3.14.0
pandas==2.1.4
openpyxl==3.1.2
python-dotenv==1.0.0
requests==2.31.0
numpy==1.26.2
django-cors-headers==4.3.1
gunicorn==21.2.0
whitenoise==6.5.0
google-generativeai==0.3.2
plotly==5.18.0